pyproject.toml
[tool.poetry]
name = "test-runner"
version = "0.6.0"
description = "A small test runner library for Python testing"
authors = ["Stephan Lukasczyk <python-test-runner@googlegroups.com>"]
license = "LGPL-3.0+"
readme = "README.md"
repository = "https://github.com/pytesting/test-runner"
keywords = "test runner, unittest, nose, pytest"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Education",
    "Topic :: Education :: Testing",
    "Topic :: Software Development",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Testing :: Unit",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = "^3.6"
virtualenv = "^16.7"
plumbum = "^1.6"
pytesting-utils = "^0.4.1"
pipfile = "^0.0.2"
attr = "^0.3.1"
deprecated = "^1.2"
benchexec = "^2.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
black = {version = "^19.10b0", allows-prereleases = true}
coverage = "^4.5"
pytest = "^5.3"
pytest-cov = "^2.8"
pylint = "^2.4"
mypy = "^0.750.0"
GitPython = "^3.0"
pytest-sugar = "^0.9.2"

[tool.black]
line-length = 88
target_version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs
    | \.git
    | \.pytest_cache
    | \.mypy_cache
    | \.venv
    | _build
    | build
    | dist
    | test-runner.egg-info
  )/
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
